add IMTreeStore
