+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/service/traverser/Traverser.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.metadata.mtree.traverser;
+package org.apache.iotdb.db.metadata.mtree.service.traverser;
 
 import org.apache.iotdb.db.exception.metadata.IllegalPathException;
 import org.apache.iotdb.db.exception.metadata.MetadataException;
