+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/service/traverser/collector/CollectorTraverser.java
@@ -16,11 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.metadata.mtree.traverser.collector;
+package org.apache.iotdb.db.metadata.mtree.service.traverser.collector;
 
 import org.apache.iotdb.db.exception.metadata.MetadataException;
 import org.apache.iotdb.db.metadata.mnode.IMNode;
-import org.apache.iotdb.db.metadata.mtree.traverser.Traverser;
+import org.apache.iotdb.db.metadata.mtree.service.traverser.Traverser;
 import org.apache.iotdb.db.metadata.path.PartialPath;
 
 // This class defines the generic resultSet as traversal result and add more restrictions on MTree
