+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/service/traverser/collector/EntityCollector.java
@@ -16,31 +16,47 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.metadata.mtree.traverser.counter;
+package org.apache.iotdb.db.metadata.mtree.service.traverser.collector;
 
 import org.apache.iotdb.db.exception.metadata.MetadataException;
+import org.apache.iotdb.db.metadata.mnode.IEntityMNode;
 import org.apache.iotdb.db.metadata.mnode.IMNode;
 import org.apache.iotdb.db.metadata.path.PartialPath;
 
-// This class implements storage group count function.
-public class StorageGroupCounter extends CounterTraverser {
+// This class defines EntityMNode as target node and defines the Entity process framework.
+public abstract class EntityCollector<T> extends CollectorTraverser<T> {
 
-  public StorageGroupCounter(IMNode startNode, PartialPath path) throws MetadataException {
+  public EntityCollector(IMNode startNode, PartialPath path) throws MetadataException {
     super(startNode, path);
+  }
+
+  public EntityCollector(IMNode startNode, PartialPath path, int limit, int offset)
+      throws MetadataException {
+    super(startNode, path, limit, offset);
   }
 
   @Override
   protected boolean processInternalMatchedMNode(IMNode node, int idx, int level) {
-    return node.isStorageGroup();
+    return false;
   }
 
   @Override
-  protected boolean processFullMatchedMNode(IMNode node, int idx, int level) {
-    if (node.isStorageGroup()) {
-      count++;
-      return true;
-    } else {
-      return false;
+  protected boolean processFullMatchedMNode(IMNode node, int idx, int level)
+      throws MetadataException {
+    if (node.isEntity()) {
+      if (hasLimit) {
+        curOffset += 1;
+        if (curOffset < offset) {
+          return true;
+        }
+      }
+      collectEntity(node.getAsEntityMNode());
+      if (hasLimit) {
+        count += 1;
+      }
     }
+    return false;
   }
+
+  protected abstract void collectEntity(IEntityMNode node) throws MetadataException;
 }
