+++ server/src/main/java/org/apache/iotdb/db/metadata/mtree/store/MTreeStoreFactory.java
@@ -1,0 +1,16 @@
+package org.apache.iotdb.db.metadata.mtree.store;
+
+public class MTreeStoreFactory {
+
+  private static IMTreeStore store;
+
+  public static IMTreeStore getMTreeStore() {
+    if (store == null) {
+      store = new MemMTreeStore();
+    }
+
+    return store;
+  }
+
+  private MTreeStoreFactory() {}
+}
